+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Added 'none_if_no_user' param in `content.get_user_fullname`.
   [sgeulette]
+- Added `setup.load_type_from_package` to reload a single type.
+  Moved `workflow.load_workflow_from_package` to `setup.load_workflow_from_package`.
+  [gbastien]
 
 0.61 (2022-07-01)
 -----------------